Rename references of ocean-web3 to squid-py
