+++ .travis.yml
@@ -16,5 +16,5 @@ deploy:
   on:
     tags: true
     branch: master
-    repo: oceanprotocol/ocean-web3
+    repo: oceanprotocol/squid-py
     python: 3.6
