+++ setup.cfg
@@ -7,7 +7,7 @@ tag = True
 search = version='{current_version}'
 replace = version='{new_version}'
 
-[bumpversion:file:ocean_web3/__init__.py]
+[bumpversion:file:squid_py/__init__.py]
 search = __version__ = '{current_version}'
 replace = __version__ = '{new_version}'
