+++ setup.py
@@ -33,13 +33,13 @@ setup(
     long_description=readme,
     long_description_content_type="text/markdown",
     include_package_data=True,
-    keywords='ocean-web3',
-    name='ocean-web3',
-    packages=find_packages(include=['ocean_web3']),
+    keywords='squid-py',
+    name='squid-py',
+    packages=find_packages(include=['squid_py']),
     setup_requires=setup_requirements,
     test_suite='tests',
     tests_require=test_requirements,
-    url='https://github.com/oceanprotocol/ocean-web3',
+    url='https://github.com/oceanprotocol/squid-py',
     version='0.1.2',
     zip_safe=False,
 )
