+++ squid_py/constants.py
