+++ tox.ini
@@ -8,7 +8,7 @@ python =
 [testenv:flake8]
 basepython = python
 deps = flake8
-commands = flake8 ocean-web3
+commands = flake8 squid-py
 
 [testenv]
 passenv = CODACY_PROJECT_TOKEN
@@ -23,7 +23,7 @@ deps =
 commands =
     {toxinidir}/scripts/deploy
     pip install -U pip
-    coverage run --source ocean_web3 -m py.test  --basetemp={envtmpdir}
+    coverage run --source squid_py -m py.test  --basetemp={envtmpdir}
     coverage report
     coverage xml
     python-codacy-coverage -r coverage.xml
